refactor: 리뷰사항 수정

+++ src/main/java/jpa/reservation/entity/Classroom.java
@@ -14,7 +14,7 @@ public class Classroom {
     private Long id;
 
     private String address;
-    private Long limitedMember;
+    private int limitedMember;
   //  private String image;
 
     @OneToOne(mappedBy="classroom", fetch = FetchType.LAZY)
